feat(hugr-py): move std extension types/ops in to `std` module
Not complete, see #1287

Quantum operations left in tests as they are not part of std extensions set (maybe they should be...?)
