+++ hugr-py/src/hugr/std/__init__.py
@@ -1,0 +1,1 @@
+"""Types and operations from the standard extension set."""
